requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gunicorn
starlette==0.36.3
google-cloud-bigquery==3.23.0
authlib==1.3.0
httpx==0.27.0
python-dotenv==1.0.1
itsdangerous==2.1.2
pydantic==2.7.1
google-cloud-storage
python-multipart
email-validator
cachetools